+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [3.5.1] - 2024-11-26
+### Fixed
+- Fixed inspector loading for service assets, below Unity 6
+- Fixed button top open in dashboard
+
 ## [3.5.0] - 2024-11-12
 
 ### Added